feat(policy): preserve values order within an attribute (#496)
Unsafe update order behavior is not yet implemented as that service does
not exist. However, DB function is stubbed and an integration test
preparing for it were added and skipped when tests run.

Trigger on create and delete of a value append and remove (respectively)
the value ID from the `values_order` column on the attribute definition
table.