chore: Divides CODEOWNERS into project captains (#446)
This pr introduces aligning the CODEOWNERS file to the newly defined
streams.

-  SDK (Go SDK)
-  CORE ( The core platform components)
-  KAS ( Key Access Server)
-  Authorization Service 
-  Policy ( Attributes, Subject Mappings, Resource Mappings)

In the current model we allow all members of the `developers` group to
approve a pull request. There are engineers, contractors and
non-engineers in that group. We have to reduce the blast radius of
someone approving a pull request that shouldn't be merged into main.