Remove 'nolint:unused' annotation and implement interface method in test
The 'nolint:unused' annotation on the 'isPolicyBody' method within the 'resourceLocator' structure has been removed. To circumvent misuse warning from linter, this method is now also implemented in 'nanotdf_test.go' file as a workaround allowing the function to be unused in the actual code but accessible in the test.