chore(ci): Updates deps after release is tagged (#887)
First, use the release please step to determine if a release was
created. If a library was released, build its list of downstream deps
(hard coded for now, could be queried with a script). Then, trigger
another job that updates those deps and builds a PR that updates them

Issue: https://github.com/opentdf/platform/issues/839