feat(core): Rewrap and Policy CRUD Audit Events (#889)
- cleans up structure of audit logging
- adds fields as discussed to the rewrap audit event
- adds policy CRUD audit events
- implements diff generation using https://github.com/wI2L/jsondiff
- closes #884

### Running Questions

Some questions I had as I worked on this, to be discussed in this PR or
on Slack. I will update the questions with their answers below when they
are answered

#### Deactivated Events

> Should deactivated actions be considered as `delete` or `update`
action types in the audit event?

After discussion with Pat, I'm leaning towards making these `update`
action types as the resource still technically exists and could be
"activated" with a small code change (although currently there are no
endpoints to support activating anything that has been deactivated)

#### Update `diff` vs Performance

> When we perform an update, we're doing a SQL update call, in order to
generate a "diff" for the audit event we'd have to load the resource
which adds another DB call. Is this something we want to do?

I confirmed with @jakedoublev that it's correct we don't have access to
the existing data at the time of update. Also confirmed with @strantalis
that we should load the existing item in order to calculate the diff
before the update as well.

### Audit Events in this PR

#### Rewrap

- [x] rewrap

#### Policy CRUD

- [x] attribute definition
  - [x] created
  - [x] updated
  - [x] updated (deactivated)
- [x] attribute value
  - [x] created
  - [x] updated
  - [x] updated (deactivated)
- [x] kas attribute definition assignment
  - [x] created
  - [x] deleted
- [x] kas attribute value assignment
  - [x] created
  - [x] deleted
- [x] subject mapping
  - [x] created
  - [x] updated
  - [x] deleted
- [x] condition sets
  - [x] created
  - [x] updated
  - [x] deleted
- [x] resource mappings
  - [x] created
  - [x] updated
  - [x] deleted
- [x] namespaces
  - [x] created
  - [x] updated
  - [x] updated (deactivated)
- [x] kas registry
  - [x] created
  - [x] updated
  - [x] deleted