refactor(core): remove extra jwt parse from authn interceptor (#1020)
Parsing a jwt is not a free operation and in `authn` we were first
parsing the token to extract the issuer. Currently this is not necessary
because we only support 1 issuer in our `authn` configuration.

This pull request removes that unnecessary parse call and moves to
passing the keyset directly into the jwt parse call.

The reason this came about is when we were profiling our service under
load it was obvious there was some slight over head doing this twice.


![image](https://github.com/opentdf/platform/assets/18211470/c3a458dc-d34f-4abf-9cf0-b6e1a340721c)