fix(authz): move opa out of startup call (#1048)
This pull request helps address
https://github.com/opentdf/platform/issues/1033 and introducing the
concept of a mode.

Currently we are starting `opa` for every service. When in practice it
is only consumed by the authorization service that is dependent on
executing the `rego` policy.

In this update, I have removed OPA and instead directly eval the Rego
policy, which can be either embedded or passed in via configuration.
Until we have clear use cases for the additional management
functionality that OPA provides, we can focus on just executing our
entitlements policy.