feat(sdk)!: Autoconfigure with grants (#1051)
Allows looking up the `grants` fields of the data attributes to
determine which KAS or set of KASes should be used to wrap a given
document key.

Basic flow:

```go
   s, err := sdk.New(/* config options */)
   s.CreateTDFContext(out, in, sdk.WithDataAttributes("http://your/attr/tag/value/here"))
```

this will look up any grants associated with the policy and use them to
generate a KAO that shares and/or splits the values across the given KAS
elements as needed.

Note that this changes several default behaviors:

1. Use `WithAutoconfigure(false)` to disable the new attribute kas grant
lookup code
2. Previously, calls to `WithDataAttribute` did no validation (would not
return errors). Now they use a regex to validate the FQN format (even
with autoconfigure disabled).
3. With Autoconfigure enabled, the encrypt method will do a lookup for
attribute Value objects from the policy service. To bypass this, either
disable autoconfigure or use the `WithDataAttributeValues` method
instead, which allows passing in pre-populated Value objects

The current 'fixtures' for grants will need to be updated to work with
this, as they currently give a variety of grants to things like
`kas.example.com` which do not exist