chore: Auto Approve Version Bumps (#1171)
Auto approves version bump PRs that satisfy the following conditions

- author is `opentdf-automation`
- PR title starts with `fix(core): Autobump`

I believe this will have the PR be approved by the GH actions bot. WE
may need to add that bot as a codeowner to the mod files or maybe
generate a token for a dedicated bot separate from opentdf-automation?