Implement PublicKeyBytes lookup and embed into NanoTDF
Added a new function lookupKidByPublicKey to the Provider in the KAS service, which looks up the key ID based on the provided public key. This function is utilized for key ID lookup in the NanoTDF rewrap method instead of the previous method. Also, included the PublicKeyBytes in the NanoTDF struct and other relevant areas. Moreover, created GetECCompressedKeyLengthFromECCMode in the EC Key Pair to extract the length of the compressed key given an ECC mode. The new approach allows us to associate public keys with their respective key IDs more efficiently.