feat(core): ability to run a set of isolated services (#1245)
This pr resolves https://github.com/opentdf/platform/issues/1033

It introduces two new config values `mode` and `sdk_config` at the root
level. It also removes the `enabled` field on a service.

Mode now takes a list which could be `all`, `core` or a specific
namespaced service for example `kas`. You can also combine modes like
`core,kas`.

When running in something other than `core` or `all` you are required to
specify a `sdk_config` object. This is because the core services will be
remote now and not over the internal IPC server.