Reduce DoS risk by limiting TDF Manifest to 10MB
This commit fixes GHSA-9m35-5v9f-8x2j by limiting TDF Manifest sizes to 10MB or smaller.  Prior to this fix, a protocol attack is possible to define a large manifest size which will result in a large memory allocation (even if the manifest transmitted is substantially smaller).

This commit also includes the fuzz testing that was used to discover this flaw.

In `reader` the Manifest load uses the function `ReadAllFileData`.  This function is documented to only be used on small files, however to accomplish that a call to `ReadFileSize` beforehand would be necessary.

Because the call to check the defined size is likely to be missed, this commit updates `ReadAllFileData` to instead accept the maximum size as an argument, which is validated before allocating the byte slice.