feat(ci): Get traffic metrics for OpenTDF repos  (#1540)
this PR runs a workflow once a week to get Github repo traffic metrics
for each of the opentdf public repos. it puts it in a gcp bucket where
bigquery can be integrated. Goal is to be able to visualize the data
over time to see views, referral sources, clones, etc. can be expanded
in the future if we want any more info.

https://github.com/opentdf/platform/actions/runs/10904129901
^ this was a successful run of the workflow.