feat(authz): Add name to entity id when retrieved from token (#1616)
resolves https://github.com/opentdf/platform/issues/1615
before: `jwtentity-0` or `jwtentity-1`
now includes more useful info in the entity name
```
       {
          "id": "jwtentity-0-clientid-tdf-entity-resolution-public",
          "client_id": "tdf-entity-resolution-public",
          "category": "CATEGORY_ENVIRONMENT"
        },
        {
          "id": "jwtentity-1-username-sample-user",
          "user_name": "sample-user",
          "category": "CATEGORY_SUBJECT"
        }
```
This is useful for parsing access decisions triggered by kas. Access
decisions only reference the entityID so, previously, it was be
difficult to know what entity it is referring to without looking through
the rest of the logs. This change should provide more context and save
debugging time.