chore(ci): Add initial CodeQL Workflow (#1655)
I want to explore using CodeQL to assist as part of our static analysis
strategy. Because this tool is available for free to open source
projects, `opentdf` is a good place to start this testing.

This PR adds an initial configuration which will scan the codebase once
a week. These results can be ignored for the time being. Instead the
security team will review the results and make sure we tune away any
noise first (or help in opening PR's to address any valid issues found).

[SEC-4162]:
https://virtru.atlassian.net/browse/SEC-4162?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ