fix(core): Set token endpoint manually if client creds provided in server sdk_config (#1780)
### Proposed Changes

We run the server sdk in IPC mode so it does not reach out to well-known
on instantiation because well-known isnt running yet. This means the
token endpoint is never set. So if a service using the server sdk wanted
to do a decrypt for example (a request that needs an auth token) they
would get an error.
This will manually set the token endpoint on instantiation using the
issuer provided in the yaml config.

### Checklist

- [ ] I have added or updated unit tests
- [ ] I have added or updated integration tests (if appropriate)
- [ ] I have added or updated documentation

### Testing Instructions