fix(sdk): Expose oauth package (#1779)
### Proposed Changes

We expose functions like NewIDPAccessTokenSource and
NewCertExchangeTokenSource and others which take ClientCredentials,
CertExchangeInfo, or TokenExchangeInfo as input -- these are currently
internal. We should expose these if we expose functions that take their
types as inputs.

If we are uncomfortable with exposing all of the oauth package, we can
break off the types and make a new public package where those can live.
Or create helper function like CreateClientCredentials in exposed
packages the results of which could be fed into the above functions.


### Checklist

- [ ] I have added or updated unit tests
- [ ] I have added or updated integration tests (if appropriate)
- [ ] I have added or updated documentation

### Testing Instructions