fix(sdk): everything is `mixedSplits` now (#1861)
### Proposed Changes

Previously we handled TDFs with missing split IDs as if each key split
with no split ID should be
part of a different split. This was incompatible with the way that
FileWatcher created TDFs; multiple
KAOs with no split IDs are part of a single split.
 
This removes this special-casing that groups splits with no IDs
separately.

Tests for this case are [in a new
xtest](https://github.com/opentdf/tests/pull/240)

### Checklist

- [ ] I have added or updated unit tests
- [ ] I have added or updated integration tests (if appropriate)
- [ ] I have added or updated documentation

### Testing Instructions