fix: Improve http.Client usage for security and performance (#1910)
### Proposed Changes
This change switches us from maintaining a tls config which we then
on-demand initialize an `http.Client` with to instead maintain and reuse
an `http.Client` instance. This enables us to utilize the connection
pooling which occurs within the `http.Transport` to reduce ssl
handshakes and thus reduce latency.

In addition this change provides us a central place to configure out
`http.Client` (`httputil`). Allowing us to easily set configuration
options to reduce the security risks of using an unconfigured
`http.Client`. Notably timeouts to reduce DoS risks, and control around
following redirects to prevent blind SSRF's.

This PR 3/4 addresses #1891. After it's merged a small change will be
made in the `service` to fully utilize the API being introduced in the
SDK here.

### Checklist

- [ ] I have added or updated unit tests
- [ ] I have added or updated integration tests (if appropriate)
- [ ] I have added or updated documentation

### Testing Instructions