fix: Service utilize `httputil.SafeHttpClient` (#1926)
### Proposed Changes

This PR follows #1910 by updating `service` to utilize the new
`httputil` helper for constructing a client which wont follow redirects,
and has sensible timeouts. To quote that description:
> This change switches us from maintaining a tls config which we then
on-demand initialize an http.Client with to instead maintain and reuse
an http.Client instance. This enables us to utilize the connection
pooling which occurs within the http.Transport to reduce ssl handshakes
and thus reduce latency.
>
> In addition this change provides us a central place to configure out
http.Client (httputil). Allowing us to easily set configuration options
to reduce the security risks of using an unconfigured http.Client.
Notably timeouts to reduce DoS risks, and control around following
redirects to prevent blind SSRF's.

The prior auth API was marked as deprecated. There is no remaining use
within this repo, so it may be able to be removed.

Merging this should fully resolve #1891.

### Checklist

- [X] I have added or updated unit tests
- [ ] I have added or updated integration tests (if appropriate)
- [ ] I have added or updated documentation

### Testing Instructions