chore(ci): Add govulncheck to checks.yaml
Previously, we were only running this in a timed job, to avoid slowing down developers if a new vulnerability is discovered. Let's see if this is really a problem. If this is causing developer pain, we can add an if block to the step that skips it (or makes it non-blocking) on pushes to main, but blocks release PRs.