fix(core): bump to go 1.24 and bump service proto module dependencies (#2064)
Buf has [new proto validation
features](https://buf.build/docs/protovalidate/schemas/predefined-rules/#example-case)
we want to take advantage of, and we need to bump the underlying service
deps in advance of utilizing those features in proto gencode. As one
module requires a minimum of `1.23.4`, we are bumping to `1.24.0` to
avoid pinning to frequently-changing patch versions.