chore(ci): add label workflow (#2003)
### Proposed Changes

Adding label workflow which has one job at this time and that is to
label pull requests with `external-contributor` if the user opening the
pull request is not a member of the org.

This workflow keys off of `author_associations` from the pull request
context.

Best
[link](https://docs.github.com/en/graphql/reference/enums#commentauthorassociation)
that I could find to describe each association

Also adds the following actions
- https://github.com/CodelyTV/pr-size-labeler
- https://github.com/actions/labeler

### Checklist

- [ ] I have added or updated unit tests
- [ ] I have added or updated integration tests (if appropriate)
- [ ] I have added or updated documentation

### Testing Instructions