fix(sdk): Prefer KID and Algorithm selection from key mappings
When key mappings (grants now with key ids!) are available,
prefer those values when possible, even if that means having multiple
mechanism types in the same TDF (e.g. RSA keys of different lengths).
This was always the intent, but apparently didn't work before.

I've added a couple of small tests,
and now will add more to the integration tests.