fix(core): return services in the order they were registered (#2733)
### Proposed Changes

So right now I have inter-PEP dependencies, so config/PEP-A for example.
PEP-A wants to set its config to the config service, for later
viewing/editing in another interface. Currently the only place available
for this is PEP-A's Init (RegisterFunc). So, right now, this is the
solution I'm using. I think there is an alternative with some
"OnServicesStarted" type hook or something as well (which might be more
useful maybe).


### Checklist

- [ ] I have added or updated unit tests
- [ ] I have added or updated integration tests (if appropriate)
- [ ] I have added or updated documentation

### Testing Instructions