fix(sdk): newGranter nil check (#2729)
This pull request makes a targeted fix to the error handling logic in
the `newGranter` function within `sdk/tdf.go`. The change ensures that
if an error occurs while creating the granter, the function immediately
returns the error, preventing further operations on a potentially
invalid object.

Error handling improvement:

* Added an early return in the `newGranter` function to immediately
return if `err` is not nil, improving reliability and preventing
assignment to `g.keyInfoFetcher` when an error occurs. (`sdk/tdf.go`,
[sdk/tdf.goR458-R460](diffhunk://#diff-04aae07cffd4333871c74efd45d47c4e6739b11bf4ebd9a9b5e90376cbf46d08R458-R460)

---------

Co-authored-by: David Mihalcik <dmihalcik@virtru.com>