fix(core): Fix mode negation and core mode (#2779)
### Proposed Changes

* There was an issue where services started in core mode were registered
but not starting. It looks like they were registered with the namespace
as the mode instead of the specified mode from the config. This PR
corrects this and adds test checks to ensure services registered are
also enabled.
* Namespaces will be registered with the mode that caused them to be
included -- if all is in the list then the namespaces will have mode
all, if core,ers is the mode then policy will have mode core and
entityresolution will have mode entityresolution ...

### Checklist

- [ ] I have added or updated unit tests
- [ ] I have added or updated integration tests (if appropriate)
- [ ] I have added or updated documentation

### Testing Instructions