feat(core): Actually use KeyManager ProviderConfig (#2837)
### Proposed Changes

- Looking up a key now loads the `manager` field in the key provider
- This is required, as the manager should be the factory key, not the
name
- This allows us to load the same manager multiple times with different
configs.

HOWEVER the config is only looked at on the first load; we should update
this so it evicts and reloads the provider if the config changes. This
hopefully will come in a follow-up.

Similarly, we don't have much in the way of integration tests for these,
since we don't include a key manager that takes a config. I'll look into
starting the Vault sample plugin back up and running.

While I'm here, since our downstream deps no longer create them, I've
removed support for the `KeyManagerFactory` that does *not* take a
context object.

### Checklist

- [x] I have added or updated unit tests
- [ ] I have added or updated integration tests (if appropriate)
- [ ] I have added or updated documentation

### Testing Instructions